items units added

+++ src/Components/Master/AccountMaster/AccountGroup/AccountGroup.js
@@ -290,9 +290,14 @@ function AccountGroup(props) {
 
                                     <button
                                       className="ml-3"
-                                      onClick={() =>
-                                        props.onDeleteAccountGroup(user.id)
-                                      }
+                                      onClick={() => {
+                                        if (
+                                          window.confirm(
+                                            "Are you sure you wish to delete this Account Group?"
+                                          )
+                                        )
+                                          props.onDeleteAccountGroup(user.id);
+                                      }}
                                     >
                                       <i
                                         className="fa fa-trash-alt "
